feat: expose visitorId via onUpdate callback status
When using SSR, in the majority of cases the visitor id will be set on the server, first creating the visitor with a call to Flagship.newVisitor(), and then storing the visitor id in a cookie on the response header for the browser to use. This visitor id then persists for subsequent requests.

In some cases a user may land on the site on a page that does not have SSR, and is entirely CSR. In this case the FlagshipProvider will internally call Flagship.newVisitor() and set a visitor id. In order to persist this value via cookies/localstorage etc from the client, the only way currently is to have a useEffect() that depends on the visitorId returned by the useFlagship() hook.

This change exposes the visitor id via the FlagshipProvider's onUpdate status object, so it can be compared against local cookies/storage and those updated if needed.